move fonts to separate folder
